rename/rearrange to match lab
